Tweaks to paint, allowing use within message and with NULL styles
Passing in a single argument NULL to paint() is treated as passing
in no paint arguments.  An additional argument normal_is_bright will
keep the text "bolded" on exit.  This needs to be manually set when
using paint *within* a message (i.e. to paint just a fraction of one)
when in interactive mode.